requirements.txt
python-telegram-bot==12.6.1
emoji==0.3.4
sqlalchemy==1.3.16
mysqlclient==1.4.6
babel==2.8.0
python-dotenv==0.13.0